add support for sha256sums.txt tests
If we are given a sha256sums.txt file as argument, then we test all the
bundles compatible with current platform listed in the sha256sums.txt.
If sha256sums.txt file is a URL, then we download it and the bundle
files that are listed in it, and check their sha256sum.