Add a script to pull from git while checking commit signature
Because we are pulling and running scripts from this repository in a
cron job, it is safer to use signed commits. Without this, someone who
compromise the git server would be able to run arbitrary scripts on our
servers.

.gitkeyring is the gpg keyring that contains the keys of people that can
sign commits.

tools/gpggit is a gpg wrapper that uses the .gitkeyring keyring file.

tools/pull is a script that sets tools/gpggit as gpg.program in the git
config, and runs git pull with the --verify-signatures option.