Don't create thumbnails on Windows
For some reason that I did not understand yet, Image::Magick does
not work on Windows, so we're not creating thumbnails if we're
on Windows. In that case, the thumbnails should be created by the
server that receives the results (this is not done yet, but should be
done later).