use tor-browser-bundle.git tags to know the latest version to test
We're now cloning builders/tor-browser-bundle.git and using the latest
signed tag on the master branch to find the current version we need to
test.

We are also using the sha256sum of the file sha256sums.txt as a build
identifier, and running the tests again when that identifier changes
(when a new build of the same version is available).