Add mozmill-tests/tbb-tests/test_page.js
This mozmill test can be used to test a page. It uses infos from the
test and options to find which page to load.

The test page should set a test_result element on the page with OK if
the test was successful, or an error otherwise.

The following test options are used:

- the name of the test is the name of the page that will be loaded (with
  .html appended)

- if remote is set to true, it is loaded remotely, using the
  test_data_url option, otherwise it is loaded from the local test-data
  directory.

- timeout is the time in milliseconds to wait for the test_result
  element to appear. The default is 50000.

- interval is how often we should check if the test_result element
  appeared. The default is 100.