Don't install CPAN version of dual-life libfile-spec-perl and libdigest-sha-perl modules on Debian.
These modules are shipped with perl, perl-base or perl-modules already, since
they are part of Perl core in all versions of Perl we care about (i.e. 5.9.3,
while even Squeeze has 5.10.1). Presumably, additionally installing the CPAN
version of these modules was a mere oversight, and the version shipped with Perl
core is good enough => let's get rid of those duplicates.

And, while I'm at it, document why the other modules that are in the same
situation, and that we _already_ did not install the CPAN version of, are
handled this way.