Bug 21982: fix svg-enable test
After the marionette and ESR52 update, the svg-disable test was broken
because of how we tried to close the download prompt window. It seems
this window is not blocking anymore, so we don't need to close it, and
instead we just check if an svg element is present inside the iframe.