Bug 25817: remove old tbb-nightly config files and scripts
In #25817 we are adding to tor-browser-build some ansible roles to set
up a Tor Browser nightly build machine. The testsuite configuration
files and scripts are now stored in the ansible roles. To avoid confusion,
we remove the old configuration files and scripts which are no longer
used when the setup is done using ansible.