Bug 40036: added a test to check that permissions are not saved on storage
In our Firefox patch set we have a test that assures that preferences
are kept in memory, instead of saving them to the disk or other
storage.
However, usually we check this kind of functionality here, not on
the Firefox/tor-browser code.

Other tests in that tor-browser commit should already be covered.