Bug 40035: Added a fonts fingerprinting test
Added a test to check that fingerprinting can only detect the fonts we
bundle with Tor Browser.
The code is readapted from TorZilla Print:
https://arkenfox.github.io/TZP/tests/fontlists.html
The list should be updated to match the bundled fonts. Also, we ship
different fonts in different platforms, so the test is enabled only for
Linux.

Added also a test to detect which glyph the browser can render and
their size. The code was originally developed by David Fifield and
Serge Egelman:
https://www.bamsoftware.com/talks/fc15-fontfp/fontfp.html#demo
The expected result of the test comes directly from a run on Tor
Browser, and we want it to be persistent, rather than having some
deisred metrics to match.